Networking

Networking

This section documents the interface and types needed to connect with a broker.

JlMQTT.Connect — Method.

JlMQTT.Connect(network::JlMQTT.MqttNetworkChannel)

Connects to a broker.

Parameters:

network - [in] JlMQTT.MqttNetworkChannel

JlMQTT.Close — Method.

JlMQTT.Close(network::JlMQTT.MqttNetworkChannel)

Closes connection to a broker.

Parameters:

network - [in] JlMQTT.MqttNetworkChannel

JlMQTT.Write — Method.

JlMQTT.Write(network::JlMQTT.MqttNetworkChannel, buffer::Vector{UInt8})

Sends buffer::Vector{UInt8} to a broker.

Parameters:

network - [in] JlMQTT.MqttNetworkChannel

buffer - [in] Vector{UInt8}

JlMQTT.Read — Method.

JlMQTT.Read(network::JlMQTT.MqttNetworkChannel, buffer::Vector{UInt8})

Receives buffer::Vector{UInt8} from a broker.

Parameters:

network - [in] JlMQTT.MqttNetworkChannel

buffer - [in] Vector{UInt8}

Returns:

[out] returns the number of read bytes

Types

JlMQTT.MqttNetworkChannel — Type.

JlMQTT.MqttNetworkChannel

Representation of a connection to a broker which provides a TCPSocket, remote address as String and remote port as Integer.