Public Interface

Public Interface

This section documents the public interface that provides an abstraction layer on MQTT messages and framing.

JlMQTT.MqttConnect — Method.

JlMQTT.MqttConnect(client::JlMQTT.MqttClient, clientId::String)

Sends CONNECT message to broker and receives CONNACK message from broker.

Parameters:

client - [in] JlMQTT.MqttClient

clientId - [in] client identifier of type String

username - [optional] type String

password - [optional] type String

will - [optional] JlMQTT.WillOptions

willFlag - [optional] will flag of type Boolean

cleanSession - [optional] clean session flag of type Boolean

Returns:

[out] return code received from broker CONNACK message

JlMQTT.MqttDisconnect — Method.

JlMQTT.MqttDisconnect(client::JlMQTT.MqttClient)

Sends DISCONNECT message to broker and closes all open channels.

Parameters:

client - [in] JlMQTT.MqttClient

Returns:

JlMQTT.MqttSubscribe — Method.

JlMQTT.MqttSubscribe(client::JlMQTT.MqttClient, topics::Vector{String}, qosLevels::Vector{UInt8})

Sends SUBSCRIBE message to broker and returns SUBSCRIBE message identifier.

Parameters:

client - [in] JlMQTT.MqttClient

topics - [in] collection of topics of type Vector{String}

qosLevels - [in] collection of QOS LEVELS of type Vector{UInt8}

Returns:

[out] SUBSCRIBE message identifier

JlMQTT.MqttUnsubscribe — Method.

JlMQTT.MqttUnsubscribe(client::JlMQTT.MqttClient, topics::Vector{String})

Sends UNSUBSCRIBE message to broker and returns UNSUBSCRIBE message identifier.

Parameters:

client - [in] JlMQTT.MqttClient

topics - [in] collection of topics of type Vector{String}

Returns:

[out] UNSUBSCRIBE message identifier

JlMQTT.MqttPublish — Method.

JlMQTT.MqttPublish(client::JlMQTT.MqttClient, topic::String, message::Vector{UInt8}; qos::JlMQTT.QosLevel = AT_MOST_ONCE, retain::Bool = false)

Sends PUBLISH message to broker and returns PUBLISH message identifier.

Parameters:

client - [in] JlMQTT.MqttClient

topic - [in] publish to topic of type Vector{String}

message - [in] message to publish of type Vector{UInt8}

qos - [optional] message QoS level

retain - [optional] retain flag

Returns:

[out] SUBSCRIBE message identifier

Types

JlMQTT.MqttClient — Type.

JlMQTT.MqttClient

Representation of a MQTT Client which provides a JlMQTT.MqttNetworkChannel to a broker.

JlMQTT.MqttClientConstructor — Function.

JlMQTT.MqttClientConstructor()

Constructor for JlMQTT.MqttClient.

Parameters:

channel - [in] JlMQTT.MqttNetworkChannel

Returns:

[out] instance of JlMQTT.MqttClient

JlMQTT.WillOptions — Type.

JlMQTT.WillOptions

Represents the Last Will and Testament information.